# Improve performance by a lot (although quite a dirty patch)