Right now nim-markdown recompiles all regular expressions on EACH call to any method/proc, so that if some proc is called 100 times, `re` module (which uses pcre) has to recompile the same regular expression 100 times - see documentation of https://nim-lang.org/docs/re.html#re%2Cstring

In the long run it would be better to actually define variables with precompiled expressions like:
```nim
let someNameForMyRe = re"abc"

proc test = 
  let a = match("helloabc", someNameForMyRe)
```

but this patch will work for now.

For a simple MD document (taken from Casa) - https://gist.github.com/Yardanico/3ed52233c29b6328bedce30a9f332c3a

I got a ~8x time improvement - from 27.5ms to 3ms:
```
# Without patch
name ............................... min time      avg time    std dv   runs
test .............................. 26.906 ms     27.842 ms    ±0.512   x179
# With patch
name ............................... min time      avg time    std dv   runs
test ............................... 2.936 ms      3.022 ms    ±0.086  x1000
```

This is already much better, but I think that we can make it even faster :)