@Yardanico 2021-03-05
Another benchmark (with https://github.com/treeform/benchy):
```nim
import benchy
import markdown

let data = "# Hello world"

timeIt "test":
  let a = markdown(data)
  keep(a)
```
With the patch:
```
name ............................... min time      avg time    std dv   runs
test ............................... 0.019 ms      0.024 ms    ±0.004  x1000
```

Without it:
```
name ............................... min time      avg time    std dv   runs
test ............................... 0.326 ms      0.345 ms    ±0.007  x1000
```

14x time improvement for a markdown hello-world!