@Yardanico 2021-03-05
Related: https://github.com/soasme/nim-markdown/issues/48