@soasme 2021-03-05
Awesome. I like the solution!